chore: vertices seq to arraybuffer + moved matrix class to separate file
